fix: stop FourBytesToInt shifting into the sign bit

+++ include/dashbls/util.hpp
@@ -143,7 +143,7 @@ class Util {
     static uint32_t FourBytesToInt(const uint8_t* bytes) {
         uint32_t sum = 0;
         for (size_t i = 0; i < 4; i++) {
-            uint32_t addend = bytes[i] << (8 * (3 - i));
+            uint32_t addend = static_cast<uint32_t>(bytes[i]) << (8 * (3 - i));
             sum += addend;
         }
         return sum;
